requirements.txt
applicationinsights==0.11.10
asgiref==3.5.2
astroid==2.11.5
backports.csv==1.0.7
boto3==1.24.2
botocore==1.27.2
certifi==2022.5.18.1
cli-helpers==0.2.3
Click==7.0
colorama==0.4.4
configobj==5.0.6
cycler==0.11.0
dill==0.3.5.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.5
django-captcha==1.0.0
django-filter==21.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-simple-captcha==0.5.17
django-storages==1.12.3
djangorestframework==3.13.1
enum34==1.1.10
filelock==3.7.1
fonttools==4.28.4
future==0.18.2
gunicorn==20.1.0
humanize==3.13.1
intel-openmp==2019.0
isort==5.10.1
jmespath==1.0.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
matplotlib==3.5.1
mccabe==0.7.0
mkl==2019.0
mssql-cli==1.0.0
numpy==1.19.3
packaging==21.3
phonenumbers==8.12.49
Pillow==8.4.0
pipenv==2022.5.2
platformdirs==2.5.2
portalocker==2.4.0
prompt-toolkit==2.0.10
psycopg2==2.8.6
Pygments==2.11.2
pyhull==2015.2.1
pylint==2.14.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0

six==1.16.0
sqlparse==0.2.4
terminaltables==3.1.10
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1